backend/requirements.txt
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
bleach==5.0.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-bleach==3.1.0
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.1.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
zope.interface==6.0
